+++ CHANGELOG.md
@@ -3,4 +3,8 @@
 ## master
 
 ## 1.0.0
+* Disable donate and premium links
 * Initial code release 
+* Remove mass processor
+* Remove Process plugin action link
+* Set plugin URL to aihr.us